pyproject.toml
[project]
name = "ZeldaGame"
version = "0.1.4"
description = "Zelda Like Game"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.7",
    "pytmx>=3.32",
]
